add migration tests
