+++ pallets/author-slot-filter/src/lib.rs
@@ -36,6 +36,11 @@ pub mod migration;
 pub mod num;
 pub mod weights;
 
+#[cfg(test)]
+mod mock;
+#[cfg(test)]
+mod tests;
+
 #[pallet]
 pub mod pallet {
 
@@ -206,95 +211,5 @@ pub mod pallet {
 	pub enum Event {
 		/// The amount of eligible authors for the filter to select has been changed.
 		EligibleUpdated(EligibilityValue),
-	}
-}
-
-#[cfg(test)]
-pub mod tests {
-	use super::*;
-	use crate as author_slot_filter;
-
-	use frame_support::{assert_ok, parameter_types, traits::Everything};
-	use frame_support_test::TestRandomness;
-	use sp_core::H256;
-	use sp_io::TestExternalities;
-	use sp_runtime::{
-		testing::Header,
-		traits::{BlakeTwo256, IdentityLookup},
-	};
-	use sp_std::vec;
-	const AUTHOR_ID: u64 = 1;
-
-	pub fn new_test_ext() -> TestExternalities {
-		let t = frame_system::GenesisConfig::default()
-			.build_storage::<Test>()
-			.unwrap();
-		TestExternalities::new(t)
-	}
-
-	type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
-	type Block = frame_system::mocking::MockBlock<Test>;
-
-	// Configure a mock runtime to test the pallet.
-	frame_support::construct_runtime!(
-		pub enum Test where
-			Block = Block,
-			NodeBlock = Block,
-			UncheckedExtrinsic = UncheckedExtrinsic,
-		{
-			System: frame_system::{Pallet, Call, Config, Storage, Event<T>},
-			AuthorSlotFilter: author_slot_filter::{Pallet, Call, Storage, Event, Config},
-		}
-	);
-
-	parameter_types! {
-		pub const BlockHashCount: u64 = 250;
-	}
-	impl frame_system::Config for Test {
-		type BaseCallFilter = Everything;
-		type BlockWeights = ();
-		type BlockLength = ();
-		type DbWeight = ();
-		type Origin = Origin;
-		type Index = u64;
-		type BlockNumber = u64;
-		type Call = Call;
-		type Hash = H256;
-		type Hashing = BlakeTwo256;
-		type AccountId = u64;
-		type Lookup = IdentityLookup<Self::AccountId>;
-		type Header = Header;
-		type Event = Event;
-		type BlockHashCount = BlockHashCount;
-		type Version = ();
-		type PalletInfo = PalletInfo;
-		type AccountData = ();
-		type OnNewAccount = ();
-		type OnKilledAccount = ();
-		type SystemWeightInfo = ();
-		type SS58Prefix = ();
-		type OnSetCode = ();
-	}
-	parameter_types! {
-		pub Authors: Vec<u64> = vec![AUTHOR_ID];
-	}
-	impl Config for Test {
-		type Event = Event;
-		type RandomnessSource = TestRandomness<Self>;
-		type PotentialAuthors = Authors;
-		type WeightInfo = ();
-	}
-
-	#[test]
-	fn set_eligibility_works() {
-		new_test_ext().execute_with(|| {
-			let value = num::NonZeroU32::new(34);
-
-			assert_ok!(AuthorSlotFilter::set_eligible(
-				Origin::root(),
-				value.clone()
-			));
-			assert_eq!(AuthorSlotFilter::eligible_count(), value)
-		});
 	}
 }
